totalSentiment() addition
The totalSentiment method finds the first word by using the already defined textToString method which puts a space in between the words. And see's if the length is greater than 0. Then isolates that particular word and get's the value associated with it through sentinemtVal. Next the method subtracts the word you just found from the length and continues this pattern to the end of the statement.